JCLOUDS-188: Include pagination parameters in Abiquo VirtualMachineTemplateOptions
Modified the Abiquo VirtualMachineTemplateOptions builder to inherit from
the BaseFilterOptionsBuilder so it can accept pagination parameters.
Also superseded the OrderBy enum by a simple String, to avoid hardcoding
the fields that are suitable to be sorted.
The OrderBy enum has been removed and the FilterOptions class moved to a
more convenient package.