[JCLOUDS-1088] fix NPE in DeploymentToNodeMetadata.apply
The NPE is caused by passing the value of CloudService.location()
to LocationPredicates.idEquals which throws a NPE when the argument
is null. However, CloudService.location() is @Nullable and is null
when the cloud service is associated with an affinity group.

The solution in this commit checks if CloudService.location() is null
and if it is, the location is obtained from the affinity group.