RateLimit module for ProfitBricks REST
Added a module that automatically deals with rate limits in the
ProfitBricks REST api and throttles the requests as needed.

The module is configured by default in the live tests, since I've
observed several failures due to rate limits when using my account, but
not added by default. Users that want jclouds to deal automatically with
this will have to register the module when creating the context.